ref: some structure changes
